+++ src/Services/BudgetCast.Gateways/BudgetCast.Gateways.Bff/AppSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace BudgetCast.Gateways.Bff;
+
+public class AppSettings
+{
+    public string IdentityApiUrl { get; set; } = string.Empty;
+
+    public string ExpensesApiUrl { get; set; } = string.Empty;
+}
