+++ src/Services/BudgetCast.Gateways/BudgetCast.Gateways.Bff/appsettings.json
@@ -1,3 +1,5 @@
 {
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "IdentityApiUrl": "https://localhost:44305",
+  "ExpensesApiUrl": "http://localhost:5234"
 }
